+++ engine/physics/boxcollider.d
@@ -49,7 +49,7 @@ class BoxCollider : Collider {
 
     private FieldState[string] fieldStates;
     
-    override ulong drawInspector(ulong offsetX, ulong offsetY, ulong panelW) {
+    override float drawInspector(float offsetX, float offsetY, float panelW) {
       auto self = this;
       return drawFields(self, fieldStates, offsetX, offsetY, panelW);
     }
