+++ engine/physics/spherecollider.d
@@ -38,7 +38,7 @@ class SphereCollider : Collider {
 
     private FieldState[string] fieldStates;
     
-    override ulong drawInspector(ulong offsetX, ulong offsetY, ulong panelW) {
+    override float drawInspector(float offsetX, float offsetY, float panelW) {
       auto self = this;
       return drawFields(self, fieldStates, offsetX, offsetY, panelW);
     }
